octoagent/apps/gateway/pyproject.toml
[project]
name = "octoagent-gateway"
version = "0.1.0"
description = "OctoAgent Gateway - FastAPI REST API + SSE"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn>=0.30,<1.0",
    "sse-starlette>=2.0,<4.0",
    "structlog>=25.1,<26.0",
    "logfire>=3.0",
    "octoagent-core",
]

[tool.uv.sources]
octoagent-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/octoagent"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]
